Backend/requirements.txt
Django==5.0.7
asgiref==3.9.1
sqlparse==0.5.3
tzdata==2025.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
django-cors-headers==4.4.0
django-filter==24.3
django-environ==0.11.2
django-storages==1.14.4
Pillow==10.4.0
psycopg==3.2.10
psycopg-binary==3.2.10
PyJWT==2.10.1
uritemplate==4.2.0
PyYAML==6.0.2
jsonschema==4.25.1
inflection==0.5.1
attrs==25.3.0
referencing==0.36.2
rpds-py==0.27.1
pytest==8.3.2
pytest-django==4.9.0
factory_boy==3.3.0
Faker==28.0.0
